fix(nvim): lsp functions
