+++ .config/nvim/init.lua
@@ -13,5 +13,6 @@ require('internal.cursorword')
 
 require('lsp.config')
 require('lsp.setup')
+require('lsp.functions')
 
 require('snippets.react')
